Move MCP into Deckle.API, available at /mcp
Eliminates the standalone Deckle.MCP service and its separate deployment
slot. MCP tools (ProjectTools, ComponentTools, DataSourceTools, FileTools)
now live in Deckle.API/McpTools/ and are served at /mcp with API key auth.

- Add ModelContextProtocol.AspNetCore to Deckle.API
- Add ApiKey auth scheme alongside existing cookie/Google OAuth
- Register "ApiKey" policy so /mcp uses API key auth exclusively,
  leaving all other endpoints on the cookie path unchanged
- Wire up AddDeckleMcpTools() in Program.cs
- Map app.MapMcp("/mcp").RequireAuthorization("ApiKey")
- Remove Deckle.MCP and Deckle.MCP.Tests projects from solution
- Migrate all MCP tests into Deckle.API.Tests (Auth/ and McpTools/)
- Add InternalsVisibleTo(Deckle.API.Tests) for McpErrors access
- Remove mcp service from AppHost, docker-compose, and CI pipeline
