+++ src/Deckle.API.Tests/McpTools/DataSourceToolsTests.cs
@@ -1,13 +1,13 @@
 using System.Security.Claims;
 using System.Text.Json;
+using Deckle.API.McpTools;
 using Deckle.Domain.Data;
 using Deckle.Domain.Entities;
-using Deckle.MCP.Tools;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Moq;
 
-namespace Deckle.MCP.Tests.Tools;
+namespace Deckle.API.Tests.McpTools;
 
 public class DataSourceToolsTests : IDisposable
 {
