+++ src/Deckle.API/Auth/ApiKeyAuthenticationHandler.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Authentication;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
 
-namespace Deckle.MCP.Auth;
+namespace Deckle.API.Auth;
 
 public class ApiKeyAuthenticationHandler(
     IOptionsMonitor<AuthenticationSchemeOptions> options,
