+++ src/Deckle.API/McpTools/ComponentTools.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using ModelContextProtocol.Server;
 
-namespace Deckle.MCP.Tools;
+namespace Deckle.API.McpTools;
 
 [McpServerToolType]
 public sealed class ComponentTools(AppDbContext db, IHttpContextAccessor httpContextAccessor)
