+++ src/Deckle.API/McpTools/FileTools.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using ModelContextProtocol.Server;
 
-namespace Deckle.MCP.Tools;
+namespace Deckle.API.McpTools;
 
 [McpServerToolType]
 public sealed class FileTools(AppDbContext db, IHttpContextAccessor httpContextAccessor)
