+++ src/Deckle.AppHost/Deckle.AppHost.csproj
@@ -16,7 +16,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Deckle.API\Deckle.API.csproj" />
-    <ProjectReference Include="..\Deckle.MCP\Deckle.MCP.csproj" />
   </ItemGroup>
 
 </Project>
