--- src/Deckle.MCP/Deckle.MCP.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <UserSecretsId>b2f3c1d4-e5a6-47b8-9c0d-1e2f3a4b5c6d</UserSecretsId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Aspire.Npgsql.EntityFrameworkCore.PostgreSQL" Version="13.1.1" />
-    <PackageReference Include="ModelContextProtocol.AspNetCore" Version="1.3.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Deckle.Domain\Deckle.Domain.csproj" />
-    <ProjectReference Include="..\Deckle.ServiceDefaults\Deckle.ServiceDefaults.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>Deckle.MCP.Tests</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-
-</Project>
