+++ src/Deckle.slnx
@@ -9,12 +9,8 @@
     <Project Path="Deckle.Email/Deckle.Email.csproj" />
     <Project Path="Deckle.Email.Abstractions/Deckle.Email.Abstractions.csproj" />
   </Folder>
-  <Folder Name="/Services/">
-    <Project Path="Deckle.MCP/Deckle.MCP.csproj" />
-  </Folder>
   <Folder Name="/Tests/">
     <Project Path="Deckle.API.Tests/Deckle.API.Tests.csproj" />
-    <Project Path="Deckle.MCP.Tests/Deckle.MCP.Tests.csproj" />
   </Folder>
   <Project Path="Deckle.AppHost/Deckle.AppHost.csproj" />
   <Project Path="Deckle.ServiceDefaults/Deckle.ServiceDefaults.csproj" />
